Fix targeted refresh duplicating external switches
Targeted refresh uses the associations to determine what records exist
already in the database and will only create new records that aren't in
that list, and, critically, will not delete records that aren't covered
by the targeted scope.

Due to the `external_distributed_virtual_switches` and
`external_distributed_virtual_lans` associations only existing in the
`Ovirt::InfraManager` subclass, and `ActsAsStiLeafClass` not including
subclasses, the associations were only looking for Ovirt classes and
returning `[]` for ems.external_distributed_virtual_switches even though
the switches did exist.

This caused a targeted refresh of a VM to duplicate switches and lans on
every refresh, and a full refresh fixing the situation because it will
delete what isn't in the current collection.